% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AB.r
\name{AndersenBroadie}
\alias{AndersenBroadie}
\title{Andersen % Broadie algorithm (martingales from continuation values)}
\usage{
AndersenBroadie(paths, payoffs, option = "put", dt, r, div, sigma, strike)
}
\arguments{
\item{paths}{The paths}

\item{payoffs}{The payoffs realized along the paths (first argument)}

\item{option}{Whether it is a "call" or "put" option}

\item{dt}{Size of time step}

\item{r}{Interest rate}

\item{div}{Dividend yield}

\item{sigma}{Volatility}

\item{strike}{The strike}
}
\value{
A lower and an upper bound for the solution to the optimal stopping problem
}
\description{
Andersen & Brodie algorithm (martingales from continuation values) for one-dimensional geometric Brownian Motion in case of a plain vanilla payoff
}
\examples{
#Andersen & Broadie algorithm
paths<-list()
payoffs<-list()
M<-10000
N<-51
dt<-0.02
r<-0.06
div<-0
sigma<-0.2
strike<-40
S0<-36
option<-"put"

#Simulating paths
paths[[1]]<- matrix(rep(c(S0), M), ncol=1, nrow=M)

for(i in 2:N)
{
  #Antithetics variates
  mm<-rnorm(M/2, mean=0, sd=1)
  mm<-rbind(mm, -mm)

  #Construct increments
  paths[[i]]<- paths[[i-1]] * exp(dt * (r - div -0.5 * sigma * sigma) + sqrt(dt) * sigma * matrix(mm,ncol=1))
}

#Payoffs of a put option
for(i in 1:N)
{
  payoffs[[i]]<- exp(-(i-1) * dt * r)*pmax(strike-paths[[i]], 0)
}

bounds<-AndersenBroadie(paths, payoffs, option, dt, r, div, sigma, strike)
low<-bounds[1]
high<-bounds[2]
print(low)
print(high)
}
\references{
Andersen, Leif, and Mark Broadie. "Primal-dual simulation algorithm for pricing multidimensional American options." Management Science 50.9 (2004): 1222-1234.

Longstaff, Francis A., and Eduardo S. Schwartz. "Valuing American options by simulation: a simple least-squares approach." The review of financial studies 14.1 (2001): 113-147.

Glasserman, Paul. Monte Carlo methods in financial engineering. Vol. 53. Springer Science & Business Media, 2013.
}
