% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmericanPut1D.r
\name{BSOption1D}
\alias{BSOption1D}
\title{BSOption1D}
\usage{
BSOption1D(N, dt, sigma, r, div, strike, S0, M, option = "call")
}
\arguments{
\item{N}{The number of time steps}

\item{dt}{The size of time steps}

\item{sigma}{The volatility}

\item{r}{The interest rate}

\item{div}{The dividend yield}

\item{strike}{The strike price}

\item{S0}{The spot}

\item{M}{The number of simulations (The number of subsimulations in the ANdersen Broadie algorithm is 1000 times smaller)}

\item{option}{The type of option ("call" of "put")}
}
\value{
A lower and an upper bound for the true value of the optimal stopping problem
}
\description{
Lower and upper bound for a one-dimensional stopping problem in case of geometric Brownian Motion and a plain vanilla payoff (i.e. call or put) as in Longstaff & Schwartz (2001).
}
\examples{
#Longstaff & Schwartz example
bounds <- BSOption1D(51, 0.02, 0.2, 0.06, 0, 40, 36, 1e5, option="put")
lower <- bounds[1]
upper <- bounds[2]
print(lower)
print(upper)
}
\references{
Andersen, Leif, and Mark Broadie. "Primal-dual simulation algorithm for pricing multidimensional American options." Management Science 50.9 (2004): 1222-1234.

Longstaff, Francis A., and Eduardo S. Schwartz. "Valuing American options by simulation: a simple least-squares approach." The review of financial studies 14.1 (2001): 113-147.

Glasserman, Paul. Monte Carlo methods in financial engineering. Vol. 53. Springer Science & Business Media, 2013.
}
