% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS.r
\name{LongstaffSchwartz}
\alias{LongstaffSchwartz}
\title{Longstaff & Schwartz algorithm}
\usage{
LongstaffSchwartz(paths, payoffs, onlyUseInTheMoneyPaths = FALSE,
  verbose = FALSE, testRegressionPaths = FALSE,
  returnContinuationValues = FALSE)
}
\arguments{
\item{paths}{The paths (i.e. trajectories) to base the stopping decision on. In general, this can be any information useful for the stopping decision.}

\item{payoffs}{The payoffs realized along the paths (first argument)}

\item{onlyUseInTheMoneyPaths}{Whether to only use in-the-money paths for regression}

\item{verbose}{Whether to print an output}

\item{testRegressionPaths}{Whether to test paths based on the new rule before preceeding to the previous time step}

\item{returnContinuationValues}{Whether to return the approximations to the continuation values as well.}
}
\value{
A lower bound to the true value of the optimal stopping problem
}
\description{
The model-free Longstaff & Schwartz algorithm for lower bounds to optimal stopping problems.
}
\examples{
#Longstaff & Schwartz algorithm
paths<-list()
payoffs<-list()
M<-10000
N<-51
dt<-0.02
r<-0.06
div<-0
sigma<-0.2
strike<-40
S0<-36

#Simulating paths
paths[[1]]<- matrix(rep(c(S0), M), ncol=1, nrow=M)

for(i in 2:N)
{
  #Antithetics variates
  mm<-rnorm(M/2, mean=0, sd=1)
  mm<-rbind(mm, -mm)

  #Construct increments
  paths[[i]]<- paths[[i-1]] * exp(dt * (r - div -0.5 * sigma * sigma) + sqrt(dt) * sigma * matrix(mm,ncol=1))
}

#Payoffs of a put option
for(i in 1:N)
{
  payoffs[[i]]<- exp(-(i-1) * dt * r)*pmax(strike-paths[[i]], 0)
}

low<-LongstaffSchwartz(paths, payoffs)
print(low)
}
\references{
Longstaff, Francis A., and Eduardo S. Schwartz. "Valuing American options by simulation: a simple least-squares approach." The review of financial studies 14.1 (2001): 113-147.
}
\author{
Fabian Dickmann
}
